move Wifi struct to wifi module
